# docs/05-api/openapi.yaml
openapi: 3.1.0
info:
  title: Rabit API
  version: 1.0.0-mvp
  description: |
    Contract-first MVP API (ADR-0010). Rules in api-guidelines.md.
    Operations marked `x-rabit-status: planned` are reserved and not implemented in MVP.
servers:
  - url: http://localhost:8080
    description: Local api (control plane)
  - url: http://localhost:8081
    description: Local media gateway (data plane)
security:
  - bearer: []
...
paths:
...
  /v1/ops/users/{user_id}/support-summary:
    parameters: [ { name: user_id, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } } ]
    get:
      tags: [ops]
      operationId: opsUserSupportSummary
      description: >-
        A user's account state for support without database access (R18): one section per module
        (account, subscription, audio, playback, library, reports) with counts, states, timestamps and
        ids only. Never contains private content: titles, notes, filenames, audio, storage keys, URLs,
        devices or identity-provider subjects (OPS-007, T30). Each view is audited with the reason.
      parameters:
        - { name: reason, in: query, required: true, schema: { type: string, minLength: 3, maxLength: 500 } }
      responses:
        '200':
          description: Support summary
          content:
            application/json:
              schema:
                type: object
                required: [user_id, generated_at, sections]
                properties:
                  user_id: { $ref: '#/components/schemas/Id' }
                  generated_at: { $ref: '#/components/schemas/Timestamp' }
                  sections:
                    type: object
                    additionalProperties: { type: object }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
  /v1/ops/jobs:
    get:
      tags: [ops]
      operationId: opsListJobs
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [queued, running, succeeded, failed, dead] } }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200': { description: Jobs (payload omitted), content: { application/json: { schema: { $ref: '#/components/schemas/JobList' } } } }
        '403': { $ref: '#/components/responses/Error' }
  /v1/ops/jobs/{job_id}/retry:
    parameters: [ { name: job_id, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } } ]
    post:
      tags: [ops]
      operationId: opsRetryJob
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, additionalProperties: false, required: [reason], properties: { reason: { $ref: '#/components/schemas/Reason' } } } } }
      responses:
        '200': { description: Requeued, content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
...
  parameters:
...
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
  responses:
    Error:
      description: Error envelope
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Id: { type: string, pattern: '^[a-z]{3}_[0-9A-HJKMNP-TV-Z]{26}$' }
...
    Timestamp: { type: string, format: date-time }
...
    Reason: { type: string, minLength: 3, maxLength: 500 }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id, retryable]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            retryable: { type: boolean }
            details: { type: object }
...
    Job:
      type: object
      required: [job_id, kind, status, attempts, max_attempts, last_error, created_at, updated_at]
      properties:
        job_id: { $ref: '#/components/schemas/Id' }
        kind: { type: string }
        status: { type: string, enum: [queued, running, succeeded, failed, dead] }
        attempts: { type: integer }
        max_attempts: { type: integer }
        last_error: { type: [string, 'null'] }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        updated_at: { $ref: '#/components/schemas/Timestamp' }
    JobList:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Job' } }